Mortality casting (age)

+++ minos/modules/mortality.py
@@ -154,7 +154,7 @@ class Mortality(Base):
             # Update their alive status, exit time, and expected life lost to the main population frame.
             dead_pop['alive'] = pd.Series('dead', index=dead_pop.index)
             dead_pop['exit_time'] = event.time
-            dead_pop['years_of_life_lost'] = self.life_expectancy(dead_pop.index) - pop.loc[dead_pop.index]['age']
+            dead_pop['years_of_life_lost'] = (self.life_expectancy(dead_pop.index) - pop.loc[dead_pop.index]['age']).astype('float64')
             self.population_view.update(dead_pop[['alive', 'exit_time', 'cause_of_death', 'years_of_life_lost']])
 
     def calculate_mortality_rate(self, index):
